refactor(task): 重构任务服务以集成任务显示服务并简化任务查询逻辑
